Fix test cluster initialization with Citus
We found that the test databases that were created, were not correctly initialized with Citus, effectively leaving it as a 'single-node'-cluster. This has been fixed and also the corresponding issues with the schema definition in the migrations. Mainly the base schema has been split up in schema creation and the rest because Citus then correctly propagates these created schemas to the worker nodes. This removed the need for constructs with 'run_command_on_worker' statements.